feat(rename): plants & plant types
