+++ src/controllers/plants/models/plant.response.body.ts
@@ -1,57 +1,26 @@
 import { ApiProperty } from '@nestjs/swagger';
-import {
-  PlantRequirementSunNeed,
-  PlantRequirementWaterNeed,
-} from '../../../entities/plants/models/requirement.entity';
-import { Country } from '../../../entities/countries/models/countries.entity';
-import { PlantPrecocity } from '../../../entities/plants/models/plant.entity';
-import { PlantCultureType } from '../../../entities/plants/models/culture.entity';
 
-export class PlantRequirementWaterResponseBody {
-  @ApiProperty({ enum: PlantRequirementWaterNeed })
-  needs: PlantRequirementWaterNeed;
+export class PlantClassificationResponseBody {
+  @ApiProperty()
+  kingdom: string;
+
+  @ApiProperty({ isArray: true })
+  clade: string[];
 
   @ApiProperty()
-  comment: string;
-}
-
-export class PlantRequirementSunResponseBody {
-  @ApiProperty({ enum: PlantRequirementSunNeed })
-  needs: PlantRequirementSunNeed;
+  order: string;
 
   @ApiProperty()
-  comment: string;
-}
-
-export class PlantRequirementResponseBody {
-  @ApiProperty({ type: PlantRequirementWaterResponseBody })
-  water: PlantRequirementWaterResponseBody;
-
-  @ApiProperty({ type: PlantRequirementSunResponseBody })
-  sun: PlantRequirementSunResponseBody;
-
-  @ApiProperty({ type: String, isArray: true })
-  floors: string[];
-}
-
-export class PlantCultureResponseBody {
-  @ApiProperty({ required: true, enum: PlantCultureType, isArray: true })
-  cultureTypes: PlantCultureType[];
+  family: string;
 
   @ApiProperty()
-  description: string;
+  genus: string;
 
-  @ApiProperty({ type: Number })
-  spacingBetweenPlants: number;
+  @ApiProperty()
+  species: string;
 
-  @ApiProperty({ type: Number, isArray: true })
-  sowingPeriod: number[];
-
-  @ApiProperty({ type: Number, isArray: true })
-  growingOnPeriod: number[];
-
-  @ApiProperty({ type: Number, isArray: true })
-  harvestPeriod: number[];
+  @ApiProperty()
+  binomialName: string;
 }
 
 export class PlantResponseBody {
@@ -59,25 +28,13 @@ export class PlantResponseBody {
   id: string;
 
   @ApiProperty()
-  plantType: string;
-
-  @ApiProperty()
-  variety: string;
-
-  @ApiProperty({ enum: Country })
-  origin: Country;
+  name: string;
 
   @ApiProperty()
   description: string;
 
-  @ApiProperty({ enum: PlantPrecocity })
-  precocity: PlantPrecocity;
-
-  @ApiProperty({ type: PlantRequirementResponseBody })
-  requirement: PlantRequirementResponseBody;
-
-  @ApiProperty({ type: PlantCultureResponseBody })
-  culture: PlantCultureResponseBody;
+  @ApiProperty({ type: PlantClassificationResponseBody })
+  classification: PlantClassificationResponseBody;
 
   @ApiProperty()
   createdBy: string;
@@ -91,5 +48,5 @@ export class PlantResponseBody {
 
 export class PlantSearchResponseBody {
   @ApiProperty({ type: PlantResponseBody, isArray: true })
-  plants: PlantResponseBody[];
+  plant: PlantResponseBody[];
 }
