--- src/entities/plant.types/plant.types.module.ts
@@ -1,7 +1,0 @@
-import { Module } from '@nestjs/common';
-import { PlanttypesService } from './plant.types.service';
-
-@Module({
-  providers: [PlanttypesService]
-})
-export class PlanttypesModule {}
