+++ src/entities/plants/plants.module.ts
@@ -1,4 +1,7 @@
 import { Module } from '@nestjs/common';
+import { PlantsService } from './plants.service';
 
-@Module({})
+@Module({
+  providers: [PlantsService],
+})
 export class PlantsModule {}
