+++ src/entities/plants/plants.service.ts
@@ -4,35 +4,28 @@ import { Plant, PlantDocument } from './models/plant.entity';
 import { InjectModel } from '@nestjs/mongoose';
 import { ConflictException } from '@nestjs/common';
 import { BaseSearchParams } from '../base.search.params';
-import { BaseEntityService, EntityFilterOption } from '../base.entity.service';
+import { BaseEntityService } from '../base.entity.service';
 
 export class PlantSearchParams extends BaseSearchParams {
-  variety?: string;
-  origin?: string;
-  precocity?: string;
-  createdBy?: string;
-  sunNeed?: string;
-  waterNeed?: string;
-  floors?: string[];
-  plantTypes?: string[];
-  cultureTypes?: string[];
-  minSpacingBetweenPlants?: number;
-  maxSpacingBetweenPlants?: number;
-  sowingPeriod?: number[];
-  growingOnPeriod?: number[];
-  harvestPeriod?: number[];
+  name?: string;
+  kingdom?: string;
+  order?: string;
+  family?: string;
+  genus?: string;
+  species?: string;
+  binomialName?: string;
 }
 
 @Injectable()
 export class PlantsService extends BaseEntityService {
-  constructor(@InjectModel(Plant.name) private plantModel: Model<PlantDocument>) {
+  constructor(@InjectModel(Plant.name) private PlantModel: Model<PlantDocument>) {
     super();
   }
 
-  async create(plant: Plant): Promise<Plant> {
+  async create(Plant: Plant): Promise<Plant> {
     try {
-      plant._id = new mongoose.Types.ObjectId();
-      return await this.plantModel.create(plant);
+      Plant._id = new mongoose.Types.ObjectId();
+      return await this.PlantModel.create(Plant);
     } catch (exception) {
       if (exception.code === 11000) {
         throw new ConflictException();
@@ -42,54 +35,19 @@ export class PlantsService extends BaseEntityService {
   }
 
   async findOneById(id: string): Promise<Plant | undefined> {
-    return await this.plantModel.findById(id);
+    return await this.PlantModel.findById(id);
   }
 
   async search(params: PlantSearchParams): Promise<Plant[]> {
-    const {
-      pagination,
-      sunNeed,
-      waterNeed,
-      floors,
-      plantTypes,
-      cultureTypes,
-      minSpacingBetweenPlants,
-      maxSpacingBetweenPlants,
-      sowingPeriod,
-      growingOnPeriod,
-      harvestPeriod,
-      ...filters
-    } = params;
-
+    const { pagination, kingdom, order, family, genus, species, binomialName, ...filters } = params;
     let findParam = this._generateFilters(filters);
-    findParam = this._generateFilter(findParam, 'requirement.water.needs', waterNeed);
-    findParam = this._generateFilter(findParam, 'requirement.sun.needs', sunNeed);
-    findParam = this._generateFilter(
-      findParam,
-      'requirement.floors',
-      floors?.map((floor) => new mongoose.Types.ObjectId(floor)),
-    );
-    findParam = this._generateFilter(
-      findParam,
-      'plantType',
-      plantTypes?.map((plantType) => new mongoose.Types.ObjectId(plantType)),
-    );
-    findParam = this._generateFilter(findParam, 'culture.cultureTypes', cultureTypes);
-    findParam = this._generateFilter(
-      findParam,
-      'culture.spacingBetweenPlants',
-      minSpacingBetweenPlants,
-      EntityFilterOption.MIN,
-    );
-    findParam = this._generateFilter(
-      findParam,
-      'culture.spacingBetweenPlants',
-      maxSpacingBetweenPlants,
-      EntityFilterOption.MAX,
-    );
-    findParam = this._generateFilter(findParam, 'culture.sowingPeriod', sowingPeriod);
-    findParam = this._generateFilter(findParam, 'culture.growingOnPeriod', growingOnPeriod);
-    findParam = this._generateFilter(findParam, 'culture.harvestPeriod', harvestPeriod);
-    return await this.plantModel.find(findParam).skip(pagination.offset).limit(pagination.limit);
+    findParam = this._generateFilter(findParam, 'classification.kingdom', kingdom);
+    findParam = this._generateFilter(findParam, 'classification.order', order);
+    findParam = this._generateFilter(findParam, 'classification.family', family);
+    findParam = this._generateFilter(findParam, 'classification.genus', genus);
+    findParam = this._generateFilter(findParam, 'classification.species', species);
+    findParam = this._generateFilter(findParam, 'classification.binomialName', binomialName);
+
+    return await this.PlantModel.find(findParam).skip(pagination.offset).limit(pagination.limit);
   }
 }
