+++ src/entities/varieties/models/requirement.entity.ts
@@ -1,0 +1,42 @@
+import { Prop } from '@nestjs/mongoose';
+import mongoose from 'mongoose';
+import { Floor } from '../../floors/models/floor.entity';
+
+export enum VarietyRequirementWaterNeed {
+  LOW = 'LOW',
+  MEDIUM = 'MEDIUM',
+  HIGH = 'HIGH',
+}
+
+export class VarietyRequirementWater {
+  @Prop({ required: true, enum: VarietyRequirementWaterNeed })
+  needs: VarietyRequirementWaterNeed;
+
+  @Prop()
+  comment: string;
+}
+
+export enum VarietyRequirementSunNeed {
+  FULLSUN = 'FULLSUN',
+  SEMISHADE = 'SEMISHADE',
+  SHADOW = 'SHADOW',
+}
+
+export class VarietyRequirementSun {
+  @Prop({ required: true, enum: VarietyRequirementSunNeed })
+  needs: VarietyRequirementSunNeed;
+
+  @Prop()
+  comment: string;
+}
+
+export class VarietyRequirement {
+  @Prop({ required: true })
+  water: VarietyRequirementWater;
+
+  @Prop({ required: true })
+  sun: VarietyRequirementSun;
+
+  @Prop({ required: true, type: [{ type: mongoose.Schema.Types.ObjectId, ref: Floor.name }] })
+  floors: mongoose.Types.ObjectId[];
+}
